Add helper method to utils
