+++ ros/lib/utils.py
@@ -4,6 +4,7 @@ import json
 from flask import jsonify, make_response
 from flask_restful import abort
 import ast as type_evaluation
+from ros.lib.models import RhAccount, System, db
 
 
 def is_valid_uuid(val):
@@ -105,3 +106,10 @@ def sort_io_dict(performance_utilization: dict):
     performance_utilization.update({**sorted_io_dict})
     del performance_utilization['io']
     return performance_utilization
+
+
+def default_queries(ident, additional_filter=None):
+    account_query = db.session.query(RhAccount.id).filter(RhAccount.account == ident['account_number']).subquery()
+    if additional_filter is None:
+        return db.session.query(System.id).filter(System.account_id.in_(account_query))
+    return db.session.query(System.id).filter(System.account_id.in_(account_query)).filter(additional_filter)
